+++ src/MonteCarloEngine.cpp
@@ -28,8 +28,8 @@ MonteCarloEngine::MonteCarloEngine(
 double MonteCarloEngine::EvaluatePayoff() {
 	// TODO difference in days: implement this for realistic _Date representations.
 	// Stubs and holidays should be handled. Remember: the basic time step considered is one day.
-	const double T = m_end_date - m_start_date;
-	const int nSteps = 1 + T;
+	const long T = m_end_date - m_start_date;
+	const long nSteps = 1 + T;
 
 	const VarianceReduction& varianceReduction = m_montecarlo_settings.m_variance_reduction;
